% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.ablines.R
\name{all_ablines}
\alias{all_ablines}
\title{Plot vertical, horizontal and diagonal ablines}
\usage{
all_ablines(h = 0, v = 0, a = 0, b = 1, col = "#666666", ...)
}
\arguments{
\item{h}{height at which the horizontal abline is plotted}

\item{v}{x position at which the vertical abline is plotted}

\item{a}{intercept for the diagonal abline}

\item{b}{slope for the diagonal abline}

\item{col}{color}

\item{...}{additional parameters passed to abline}
}
\description{
Plot vertical, horizontal and diagonal ablines
}
\details{
Plot vertical, horizontal and diagonal ablines
}
\examples{
plot(-10:10, -10:10)
all_ablines()
}
