% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayplot.R
\name{spatial.avg}
\alias{spatial.avg}
\alias{plot.arrayplot}
\alias{arrayplot}
\title{Show an image based on row/column information}
\usage{
spatial.avg(mat, window = 20)

\method{plot}{arrayplot}(x, ...)

arrayplot(rows, cols, vals, byrow = TRUE, n.row = max(rows),
  n.col = max(cols), pal = NULL, bg.col = "min")
}
\arguments{
\item{mat}{an object of class arrayplot}

\item{window}{window size to average}

\item{rows}{integer vector with row positions}

\item{cols}{integer vector with column positions}

\item{vals}{numeric vector with values to plot}

\item{byrow}{should the image be plotted row-wise (default:TRUE)}

\item{n.row}{number of rows (if NULL, maximum value of rows)}

\item{pal}{Character vector to use as color palette for the image}

\item{bg.col}{how the missing values should be extrapolated}
}
\value{
\code{spatial.avg} and \code{arrayplot} both return an object of class arrayplot.
}
\description{
Show an image based on row/column information
}
\details{
The image matrix is created from three columns containing the value to
plot and its position (row and column) on the image.

\code{spatial.avg} can be used to average the image.
}
