% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{auc.pval}
\alias{auc.pval}
\title{shorthand to return the the p-value of AUC calculated by pROC}
\usage{
auc.pval(x, only.pval = TRUE)
}
\arguments{
\item{x}{list or data frame containing columns "reality" and "predictor"}

\item{only.pval}{if FALSE, returns the whole result of wilcox.test}
}
\description{
run the Wilcoxon test on predictions vs reality
}
