% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{auc}
\alias{auc}
\title{shorthand to return the AUC calculated by pROC}
\usage{
auc(x)
}
\arguments{
\item{x}{list or data frame containing columns "reality" and "predictor"}
}
\value{
a named vector containing AUC, confidence intervals and a p-value
}
\description{
shorthand to return the AUC calculated by pROC
}
