% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanceset.R, R/rf.R
\name{balanceset}
\alias{balanceset}
\title{Balance samples}
\usage{
balanceset(x)

balanceset(x)
}
\arguments{
\item{x}{a factor}

\item{n}{optionally, how many samples should be in each group}

\item{replace}{whether sampling with replacement}

\item{x}{a vector}
}
\description{
Balance samples

given a factor, sample the overrepresented groups such that each factor
level is represented the same number of times
}
\details{
Given a factor, sample the overrepresented groups such that each factor
level is represented the same number of times.
The function counts the number of occurences of the factor levels, and
resamples each group with number of occurences higher than minimal.
}
