% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse}
\alias{browse}
\title{Browse a large object}
\usage{
browse(x, limit.width = TRUE)
}
\arguments{
\item{x}{object to display}

\item{limit.width}{whether use the current width option}
}
\description{
Browse a large object
}
\details{
Browse a large object
}
