% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDif.R
\name{calcDif}
\alias{calcDif}
\title{Create a "difference" gene expression set}
\usage{
calcDif(E, pairID, groupID)
}
\arguments{
\item{E}{A limma \code{EList} object.}

\item{pairID}{A factor used to match the samples. Incomplete pairs are allowed, but they will be ignored.}

\item{groupID}{A factor categorizing the samples in one of two groups.  This factor must have exactly two levels.}
}
\value{
\code{EList} object with as many samples as there are pairs
present in both groups. In the \code{target} slot of the object, a new
column called \code{difID} is created (if absent) or modified, and includes
information about the subtraction.
}
\description{
This function takes a \code{limma} \code{EList} object and calculates the
   pairwise differences between two groups of samples.
}
\details{
First, the function construct a list of pair IDs which are present in
both groups. Next, for each sample in the first group, the corresponding
sample from the second group is subtracted from it. This is of course
done for every variable (gene or otherwise) in \code{E}.
}
