% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catf.R
\name{catf}
\alias{catf}
\alias{printf}
\title{Shortcut with sprintf}
\usage{
catf(...)

printf(...)
}
\arguments{
\item{...}{arguments passed to sprintf before calling the actual function}
}
\description{
Shortcut with sprintf
}
\details{
These functions are wrappers around \code{function( sprintf( ... ) )}
}
\examples{
catf("pi=\%.5f", pi)
printf("pi=\%.5f", pi)
}
