% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript.R
\name{cc}
\alias{cc}
\title{Smart counters for rmarkdown}
\usage{
cc(label)
}
\arguments{
\item{label}{character string consisting of a prefix, `_` (underscore) and actual label, e.g. `fig_label1`}
}
\description{
Smart counters for rmarkdown
}
