% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centers.predict.R
\name{predict.fclust}
\alias{predict.fclust}
\alias{predict.kmeans}
\alias{centers.predict}
\title{useful for predicting cluster membership from cmeans and kmeans}
\usage{
\method{predict}{fclust}(cl, x)

\method{predict}{fclust}(cl, x)

centers.predict(centers, x)
}
\description{
useful for predicting cluster membership from cmeans and kmeans
}
\details{
useful for predicting cluster membership from cmeans and kmeans
}
