% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustCompPlot.R
\name{clustCompPlot}
\alias{clustCompPlot}
\title{compare different clusters in an overview plot}
\usage{
clustCompPlot(c1, c2, col = "black", ...)
}
\arguments{
\item{c1, c2}{alternate cluster memberships}

\item{col}{color of the plot}

\item{...}{additional parameters passed to the \code{plot} function}
}
\description{
compare different clusters in an overview plot
}
\details{
compare different clusters in an overview plot
}
