% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustplot.R
\name{clustplot}
\alias{clustplot}
\title{first attempt at pca/clustering plot (not working?)}
\usage{
clustplot(d, pca.comp = 1, line.width = 1e-04, col = "#33333333")
}
\description{
first attempt at pca/clustering plot (not working?)
}
