% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topTableAll.R
\name{combineTopTableAll}
\alias{combineTopTableAll}
\title{Convert a topTableAll results by gene ID}
\usage{
combineTopTableAll(tt, ID, contrasts = NULL, method = "FDR")
}
\description{
Convert a topTableAll results by gene ID
}
\details{
Convert a topTableAll results by gene ID by averaging log fold changes
and calculating combined p-values
}
