% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confuMat.R
\name{confuMat}
\alias{confuMat}
\title{Calculate a confusion matrix}
\usage{
confuMat(reality, predictions, as.text = FALSE)
}
\arguments{
\item{reality}{vector with actual assignments of groups to items}

\item{predictions}{vector with predicted assignments of groups to items}

\item{as.text}{if TRUE, return a single character value; if FALSE, return a matrix}
}
\value{
\code{confuMat} invisibly returns the confusion matrix
}
\description{
Calculate a confusion matrix
}
\details{
Calculate a confusion matrix
}
