% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyan.R
\name{copyan}
\alias{copyan}
\title{Annotated PDF copy of the graph produced}
\usage{
copyan(file = "tmp.pdf", cex = 0.8, out.type = "cairo", ...)
}
\arguments{
\item{file}{File to write the copy to (default="tmp.pdf")}

\item{cex}{Scaling used to annotate the figure}

\item{out.type}{Output type passed to \code{\link{dev.copy2pdf}}}

\item{...}{additional parameters passed to \code{\link{dev.copy2pdf}}}
}
\description{
Annotated PDF copy of the graph produced
}
\details{
Produces a PDF copy which includes the information about date and
working directory. For documentation purposes.
}
