% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ding.R
\name{ding}
\alias{ding}
\title{Show a zenity dialog "Calculation complete"}
\usage{
ding(text = "Calculation complete")
}
\arguments{
\item{text}{Text to show in the dialog}
}
\description{
Show a zenity dialog "Calculation complete"
}
\details{
Show a zenity dialog "Calculation complete"
}
