% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{distVI}
\alias{distVI}
\title{Calculate the variation of information distances}
\usage{
distVI(x, discretize = TRUE)
}
\arguments{
\item{x}{matrix or data frame. Variables should be in rows}
}
\description{
Calculate the variation of information distances, based on mutal
information
}
