% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{clean.graph}
\alias{clean.graph}
\alias{dmtx2graph}
\title{Convert a distance matrix to an undirected graph}
\usage{
clean.graph(x, cutoff = NULL)

dmtx2graph(x, nodenames = NULL, data = NULL, edgemode = "undirected",
  cutoff = Inf, init.RCytoscape = TRUE, clean = TRUE)
}
\arguments{
\item{x}{distance matrix}
}
\description{
Convert a distance matrix to an undirected graph
}
