% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotpair.R
\name{dotpair}
\alias{dotpair}
\title{A variant of the pairs plot}
\usage{
dotpair(x, xlim = NULL, ylim = NULL, main = NULL,
  cor.method = NULL, ...)
}
\arguments{
\item{x}{data frame or matrix}

\item{xlim, ylim}{ranges for the plot}

\item{main}{Title of the plot}

\item{...}{additional parameters passed to the \code{plot} function}
}
\description{
A variant of the pairs plot
}
\details{
A wrapper for the pairs plot which shows a plot in the upper panel and
correlation coefficient with p-value in the lower panel
}
