% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorsummary.R
\name{factorsummary}
\alias{factorsummary}
\title{Convert columns of a data.frame to factors and show summary}
\usage{
factorsummary(data, filter = TRUE)
}
\arguments{
\item{data}{a data frame}

\item{filter}{Whether boring columns should be omitted.}
}
\description{
Convert columns of a data.frame to factors and show summary
}
\details{
Shortcut to convert the columns of a data.frame to factors and then run
the summary function on the resulting data frame. By default, columns
which either have as many unique values as there are rows, or only one
unique value are filtered.
}
