% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastboxplot.R
\name{fastboxplot}
\alias{fastboxplot}
\title{A boxplot consisting of lines only}
\usage{
fastboxplot(x, g = NULL, columns = 1, log = FALSE, thinline = 0.1,
  thickline = 1, ylab = "", xlab = "", ...)
}
\arguments{
\item{x}{a matrix or vector; if it is a matrix, each box plot will correspond to a column of this matrix}

\item{g}{grouping of values (by default, 'col(x)')}

\item{columns}{number of columns on the plot}

\item{log}{whether the scale should be logarithmic}

\item{thinline}{thickness of the thin line (range)}

\item{thickline}{thickness of the thick line (.25 - .75 quantile)}

\item{xlab, ylab}{labels for the x and y axes}
}
\description{
A boxplot consisting of lines only for fast drawing
}
\details{
Quick to draw and hundreds of boxes can be fit. I use it to visualize
microarray distributions.
}
