% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_label.R
\name{fig_label}
\alias{fig_label}
\title{Create a label at the top left of the graphics device}
\usage{
fig_label(text, region = "figure", pos = "topleft", cex = NULL, ...)
}
\arguments{
\item{text}{a character string to be plotted}

\item{region}{in which region of the device the label should be placed}

\item{pos}{position relative to the region ("topleft" by default)}

\item{cex}{magnification of the characters. See \code{\link{par}}}

\item{...}{any further parameters (e.g. \code{col}) will be passed to
the \code{\link{text}} function.}
}
\description{
Puts a text anywhere on the device, with a logical position (such as
"topleft" or "center") specified relative to one of three plotting
regions (\code{figure} (default), \code{device} or \code{plot})
}
\details{
The main purpose of this function is to add labels (such as "A", "B",
...) to figures
}
\section{Warnings}{

\code{fig_label} is not compatible with \code{ggplot2}. However,
par(mfrow=...) and layout() work just fine.
}

\examples{
## Basic use:

par(mfrow=c(2,2))
for(x in LETTERS[1:4]) { 
   plot(rnorm(100))
   fig_label(x, cex=2) 
}

## Plotting at different positions and in different regions:

plot(rnorm(100))
for(i in c("topleft", "topright", "top", 
      "left", "center", "right", 
      "bottomleft", "bottom", "bottomright")) {
      fig_label(i, pos=i, cex=2, col="blue")
      fig_label(i, pos=i, cex=1.5, col="red", region="plot")
}

## All the different regions:

par(mfrow=c(2,2))
for(x in LETTERS[1:4]) { 
   plot(rnorm(100))
   fig_label("figure region", cex=2, col="red") 
   fig_label("plot region", region="plot", cex=2, col="blue")
}
fig_label("device region", cex=2, pos="bottomright", 
  col="darkgreen", region="device")
}
\seealso{
\code{\link{layout}}, \code{\link{par}}
}
