% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genePairs.R
\name{genePairs}
\alias{genePairs}
\alias{pairsxy}
\title{Pairs plot using smoothScatter}
\usage{
genePairs(x, y = NULL, controls = NULL, xlim = NULL, ylim = NULL,
  cor.method = "pearson", ...)

pairsxy(x, y, panelfunc = NULL, xlim = NULL, ylim = NULL)
}
\description{
Pairs plot using smoothScatter
}
\details{
Pairs plot using smoothScatter
}
