% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGOrilla.R
\name{print.GORilla}
\alias{print.GORilla}
\alias{getGOrilla}
\title{Wrapper to run the online GO service GOrilla}
\usage{
\method{print}{GORilla}(g)

getGOrilla(x, y = NULL, ontology = "proc", species = "HOMO_SAPIENS",
  pval.thresh = 0.001, download = TRUE)
}
\arguments{
\item{x}{character vector with sorted gene identifiers or (in two-list mode)
the identifiers of differentially expressed genes}

\item{y}{if in two list mode: character vector with gene identifiers of the background gene set}

\item{ontology}{Ontology to use (proc, func or comp. Default: proc)}

\item{species}{Can be HOMO_SAPIENS, 
ARABIDOPSIS_THALIANA,
SACCHAROMYCES_CEREVISIAE,
CAENORHABDITIS_ELEGANS,
DROSOPHILA_MELANOGASTER,
DANIO_RERIO,
HOMO_SAPIENS,
MUS_MUSCULUS or
RATTUS_NORVEGICUS.}

\item{pval.thresh}{The p-value threshold parameter. Allowed values: 1e-3, 1e-4, ..., 1e-11}

\item{download}{Should \code{getGOrilla} attempt to download the data tables.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
An object of the class GORilla
}
\description{
Wrapper to run the online GO service GOrilla
}
\details{
Wrapper to run the online GO service GOrilla
}
