% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsubsets.R
\name{getsubsets}
\alias{getsubsets}
\title{Sample a number of samples by factor levels}
\usage{
getsubsets(f, n = 100, sel = NULL, verbose = FALSE)
}
\arguments{
\item{f}{a factor}

\item{n}{sample size. If n < 1, then n will be treated as fraction rather than number}

\item{sel}{which factor levels to choose from. Default: NULL, meaning every factor level}
}
\description{
Sample a number of samples by factor levels
}
\details{
Sample a number of samples by factor levels.
}
