% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep.df.R
\name{grep.df}
\alias{grep.df}
\title{Quick grep through all columns of a data frame}
\usage{
grep.df(arg, df, multiple = FALSE, return.rowname = TRUE, ...)
}
\arguments{
\item{arg}{The regular expression to be searched for}

\item{df}{Data frame to search}

\item{multiple}{Whether more than one result should be returned}

\item{return.rowname}{Whether rowname rather than the index should be returned}

\item{...}{further arguments passed to \code{grep}}
}
\description{
Quick grep through all columns of a data frame
}
\details{
Quick grep through all columns of a data frame
}
