% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairedswarm.R
\name{linepairs}
\alias{linepairs}
\title{Beeswarm with lines linking pairs}
\usage{
linepairs(x, y, pairs, ...)
}
\arguments{
\item{x}{x positions of the points}

\item{y}{y positions of the points}

\item{pairs}{factor defining the pairs}
}
\description{
Beeswarm with lines linking pairs
}
\details{
Beeswarm with lines linking pairs
}
