% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineq.R
\name{lineq.above}
\alias{lineq.above}
\alias{lineq}
\title{Parameters of a line drawn on a plot}
\usage{
lineq.above(lineq, m)

lineq()
}
\arguments{
\item{lineq}{List returned by \code{lineq}}

\item{m}{Matrix with two columns giving the coordinates of points on a
plot}
}
\value{
\code{lineq} returns a list with intercept and coefficient of
the line drawn between the two clicks of the user. \code{lineq.above}
returns a logical vector indicating the points
that lie above that line.
}
\description{
Parameters of a line drawn on a plot
}
\details{
Interactively draw a line on the plot and return the parameters of that
line. Useful for determining the points that are above or below the line

\code{lineq.above} takes the output of \code{lineq} and coordinates of
some points on the plot and returns a logical vector indicating the points
that lie above that line
}
