% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttmake.R
\name{mkdesign}
\alias{mkdesign}
\title{Quickly make a design}
\usage{
mkdesign(g)
}
\arguments{
\item{g}{factor to be plotted. If g is not a factor, it will be
converted to a factor.}
}
\value{
a model matrix with column names corresponding to factor levels.
}
\description{
Quickly make a model matrix from a factor
}
\details{
This is a shortcut wrapper around \code{model.matrix(~0 + g)}
}
