% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topTableAll.R
\name{MSD}
\alias{MSD}
\title{Calculate MSD based on logFC and CI}
\usage{
MSD(lfc, ci.L, ci.R)
}
\arguments{
\item{lfc}{vector of log fold changes or signs of log fold changes}

\item{ci.L}{Left boundary of confidence interval}

\item{ci.R}{Right boundary of confidence interval}
}
\value{
a vector of MSD values
}
\description{
Calculate MSD based on logFC and CI
}
\details{
MSD ("minimal significant difference") is defined as minimum between the
confidence interval boundaries multiplied by the sign of the change. MSD is
the absolute distance between the confidence interval and 0. Or, MSD is the
smaller of absolute values of confidence intervals.
}
