% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{mybarplot}
\alias{mybarplot}
\title{A variant box and whiskers plot}
\usage{
mybarplot(data, groups = NULL, col = NULL, whiskers = NULL,
  ylim = NULL, ...)
}
\description{
A variant box and whiskers plot
}
\details{
A variant box and whiskers plot
}
