% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfaces.R
\name{myfaces}
\alias{myfaces}
\title{Chernoff faces}
\usage{
myfaces(xy = rbind(1:3, 5:3, 3:5, 5:7), which.row = NULL,
  fill = FALSE, nrow, ncol, scale = TRUE, byrow = FALSE, main,
  labels, colors = "black")
}
\arguments{
\item{which.row}{row selection from the matrix}

\item{nrow, ncol}{in how many rows / columns}

\item{byrow}{samples are in columns rather than rows}

\item{colors}{colors to use for drawing}
}
\description{
Draw Chernoff faces -- replacement for the faces function from package TeachingDemos
}
\details{
essentially the same as "faces" from the R package TeachingDemos, but
with additional options controlling the output
}
