% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{mykfold}
\alias{mykfold}
\title{kfold cross-validation with additional features}
\usage{
mykfold(x, response, grouping_var = NULL, folds = 10)
}
\arguments{
\item{x}{data matrix (rows=variables, cols=samples)}

\item{response}{response (categorical)}

\item{grouping_var}{for LGOCV (leave one group out CV)}

\item{folds}{number of folds}
}
\description{
kfold cross-validation with additional features
}
