% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myloess.sd.R
\name{myloess.sd}
\alias{myloess.sd}
\title{Local Polynomial Regression Fitting with Variability bands}
\usage{
myloess.sd(x, y = NULL, level = 0.95, ...)
}
\description{
Local Polynomial Regression Fitting with Variability bands
}
\details{
Slightly modified version from the msir package. See the manual for
loess.sd in the msir package for help.
}
