% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myloo.R
\name{myloo}
\alias{myloo}
\title{leave one out cross-validation with additional functions}
\usage{
myloo(rg, grouping_var = NULL, secondary_model = FALSE, nvar = 20,
  balancing = TRUE, mask = NULL, bestsample_sel = FALSE,
  response = "GROUP")
}
\arguments{
\item{rg}{EList object from limma}

\item{grouping_var}{grouping variable}

\item{secondary_model}{whether secondary model should be calculated}

\item{nvar}{number of variables for the secondary model}

\item{balancing}{whether sets should be balanced}

\item{mask}{additional mask}

\item{bestsample_sel}{choose only samples which were predicted 
correctly for prediction of the unknown sample}

\item{response}{name of the response variable}
}
\description{
leave one out cross-validation with additional functions
}
\details{
leave one out cross-validation with additional functions
}
