% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPlotDensities.R
\name{myPlotDensities}
\alias{myPlotDensities}
\title{Alternative to plotDensities from the limma package.}
\usage{
myPlotDensities(O, group = NULL, log = F, labels = NULL,
  col = NULL, xlim = NULL, ylim = NULL, lwd = 1, ...)
}
\arguments{
\item{O}{object (e.g. limma) for plotting the densities}

\item{group}{Color the densities by group}

\item{log}{Logarithmize the values first}

\item{labels}{Labels to add at the density maxima}

\item{col}{character vector specyfing the color(s) to use for plotting}

\item{xlim, ylim}{ranges for the plot}

\item{lwd}{line width to plot with}

\item{...}{additional parameters passed to the \code{plot} command}
}
\description{
Alternative to plotDensities from the limma package.
}
\details{
Alternative to plotDensities from the limma package.
}
