% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nalomf.R
\name{na.lomf}
\alias{na.lomf}
\title{Last observation moved forward}
\usage{
na.lomf(x)
}
\description{
Replaces all NA values with last non-NA value
}
\details{
From http://stackoverflow.com/questions/7735647/replacing-nas-with-latest-non-na-value
}
