% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{os}
\alias{os}
\title{Report memory usage}
\usage{
os(n = 10, x = NULL)
}
\arguments{
\item{n}{Number of largest objects to show in output}

\item{x}{Name of the object for which to show the output}
}
\description{
Report memory usage
}
\details{
Report memory usage for the largest objects defined.
}
