% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairedLOO.R
\name{pairedLOO}
\alias{pairedLOO}
\title{RF LOO removing whole group of subjects}
\usage{
pairedLOO(data, class, group)
}
\arguments{
\item{data}{Variables for machine learning}

\item{class}{Response variable (two classes task)}

\item{group}{factor defining how the samples are grouped}
}
\description{
RF LOO removing whole group of subjects
}
\details{
RF LOO removing whole group of subjects. For each unique value G of group,
the training set consists of all samples for which group is different from
G, and the test set contains only the samples for which group is equal G
}
