% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairedswarm.R
\name{pairedswarm}
\alias{pairedswarm}
\title{Beeswarm with lines linking pairs}
\usage{
pairedswarm(y, g, pairs = NULL, pwcol = NULL, lcol = "grey", ...)
}
\arguments{
\item{y}{values to be plotted}

\item{g}{group assignments}

\item{pairs}{vector of length equal to number of data points plotted,
indicating the pair ID for each data point}

\item{...}{Any further parameters to pass to \code{beeswarm}}

\item{color}{of the lines to draw}
}
\description{
Beeswarm with lines linking pairs
}
\details{
Beeswarm with lines linking pairs.
}
