% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca2var.R
\name{plotPcaVar}
\alias{plotPcaVar}
\alias{pca2var}
\title{Calculate the variance explained for prcomp objects}
\usage{
plotPcaVar(p2v, add = FALSE, ...)

pca2var(pca, cumulative = FALSE)
}
\arguments{
\item{p2v}{object returned by \code{pca2var}}

\item{pca}{a prcomp object}

\item{cumulative}{whether a cumulative fraction of variance should be calculated}
}
\description{
Calculate the variance explained for prcomp objects
}
\details{
Calculate the variance explained for prcomp objects
}
