% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaR2plot.R
\name{pcaR2}
\alias{pcaR2}
\alias{plot.pcaR2}
\title{Calculate R-squared for a matrix or principal components}
\usage{
pcaR2(x, predictor, randomize = 0, comps = NULL, scaled = FALSE)

\method{plot}{pcaR2}(x, ...)
}
\arguments{
\item{x}{a numeric matrix or object of class prcomp}

\item{predictor}{a factor or a numeric vector}

\item{randomize}{how many randomizations should be run}

\item{comps}{for which PCA components (or for which columns of the numeric matrix) the correlations should be calculated}

\item{scaled}{whether the data should be scaled}
}
\description{
Calculate the fraction of variance explained by a predictor (numerical
or factor) for each of the columns in a matrix or for each principal
component in a prcomp object.
}
\details{
Calculate the fraction of variance explained by a predictor (numerical
or factor) for each of the columns in a matrix or for each principal
component in a prcomp object.
}
