% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaR2plot.R
\name{pcaR2plot}
\alias{pcaR2plot}
\title{Automatic R2 plot}
\usage{
pcaR2plot(pca, df, selvars = colnames(df), comps = NULL,
  scaled = FALSE, col = NULL, ...)
}
\arguments{
\item{pca}{prcomp object}

\item{df}{data frame containing predictors}

\item{selvars}{names of the columns from data frame to choose as predictors}

\item{comps}{for which PCA components (or for which columns of the numeric matrix) the correlations should be plotted}

\item{scaled}{whether the data should be scaled}

\item{col}{colors to use for the different predictors}

\item{...}{additional parameters to be passed to plotting functions such as pch or lty}
}
\description{
Plots fraction of explained variance for PCA components and 
some additonal explanatory variables
}
\details{
Plots fraction of explained variance for PCA components and 
some additonal explanatory variables
}
