% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenes.R
\name{plotGenes}
\alias{plotGenes}
\title{Simple function to plot gene expression}
\usage{
plotGenes(x, center = TRUE, scale = TRUE, labels = NULL,
  xval = NULL, sel = NULL, aveby = NULL, xlim = NULL,
  ylim = NULL, col = "#33333366", sdcol = NULL, lwd = 1,
  add = FALSE, las = 2, genelabels = rownames(x), debug = FALSE,
  xlab = "", ylab = "", hline = 0, vlines = T, ...)
}
\arguments{
\item{x}{A matrix containing gene expression data (row-wise: genes; col-wise: samples)}

\item{center}{Whether gene expressions should be centered to mean= 0}

\item{scale}{Whether gene expressions should be scaled to sd= 1}

\item{labels}{Labels for the x-axis}

\item{xval}{Values for the x-axis}

\item{sel}{Selection of rows from the matrix to display}

\item{aveby}{If there are replicates that should be averaged before
plotting, this vector should contain IDs unique for a batch, but identical
between replicates in a batch, of length equal to number of columns in
\var{x}}

\item{xlim, ylim}{limits for x- and y- axes}

\item{col}{default color to use in plotting}

\item{sdcol}{color used to plot the standard deviations if \var{aveby}
is not NULL}

\item{lwd}{default line width to use in plotting}

\item{add}{Whether the existing plot should be updated rather than
replaced}

\item{las}{use las=2 to make sure x-axis labels are vertical}

\item{genelabels}{labels shown on the left for each gene plotted. Must
have a length equal to number of rows in \var{x}.}

\item{debug}{Switch on debugging}

\item{xlab}{Label for the x axis}

\item{ylab}{Label for the y axis}

\item{hline}{Whether a grey line should be drawn at h=0}

\item{vlines}{Whether grey vertical lines should be drawn at each sample}

\item{...}{Additional parameters will be passed to \code{plot}}
}
\description{
Simple function to plot individual gene expression for a number of genes
}
