% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myloess.sd.R
\name{plotMyloess}
\alias{plotMyloess}
\title{Plot an object from myloess.sd}
\usage{
plotMyloess(lo, add = T, col = "#990066", pred.int = TRUE,
  conf.int = TRUE, ...)
}
\description{
Plot an object from myloess.sd
}
