% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocplot.R
\name{plotROCseries}
\alias{plotROCseries}
\title{Plot a series of ROCs}
\usage{
plotROCseries(prs, legend = TRUE, legend.names = NULL,
  add.stats = TRUE, col = NULL, lwd = 3, lty = NULL, ...)
}
\arguments{
\item{prs}{a list returned by prconvert}

\item{legend}{whether a legend should be plotted on the bottom right of the figure}

\item{legend.names}{a vector of names to plot of the legend}

\item{add.stats}{whether AUC and CI should be added to the legend}

\item{col}{vector of colors of the length equal to length of prs}

\item{lwd}{line width}

\item{lty}{line type}

\item{...}{any further arguments are passed to the rocplot function}
}
\value{
A list of objects returned by the rocplot function.
}
\description{
Plot a series of ROCs
}
\details{
Plot a series of receiver-operator curves (ROCs). This function is a wrapper around rocplot(),
allowing to plot multiple ROCs.

ROCs to be drawn are provided in the prs parameter, a list, for which
each element is a data frame containing at least the following columns:
reality, decision and prediction. The column reality describes the actual
(real) classification; "decision" is the class assigned by the ML
algorithm, and "prediction" is a numeric value describing the probability
of a given sample to belong to the case class.

The graphic parameters col, lwd and lty can be either a single value,
which will be applied to all lines drawn, or a vector of the same length as
the number of ROCs to be drawn.
}
\seealso{
rocplot, rocformat, prconvert
}
