% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcomp2.R
\name{prcomp2}
\alias{prcomp2}
\title{A wrapper around prcomp}
\usage{
prcomp2(x, removeZeroVariance = TRUE, transpose = TRUE, scale = TRUE,
  ...)
}
\description{
A wrapper around prcomp
}
\details{
This function first transposes the matrix and then eliminates zero variance rows.
}
