% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocplot.R
\name{prconvert}
\alias{prconvert}
\title{Convert RF votes to plottable object}
\usage{
prconvert(pr, reality, caseclass = 1, classnames = unique(reality))
}
\value{
an object plottable with the rocplot() function
}
\description{
Convert RF votes to plottable object
}
