% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{qval2signedqval}
\alias{qval2signedqval}
\title{Change q values to logarithmized signed q values}
\usage{
qval2signedqval(qval, lfc, log = TRUE)
}
\arguments{
\item{qval}{a vector of q values}

\item{lfc}{a vector of directions of change (e.g. log fold changes)}

\item{log}{whether the q values should be logarithmized}
}
\description{
Change q values to logarithmized signed q values
}
