% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repsapply.R
\name{repsapply}
\alias{repsapply}
\title{A more general version of avereps}
\usage{
repsapply(x, ID, colwise = TRUE, func = mean)
}
\arguments{
\item{x}{a numeric matrix}

\item{ID}{function will be applied to replicates with the same ID}

\item{colwise}{If TRUE, then for each column, the function will be
applied to rows with identical IDs. If FALSE, then for each row, the
function will be applied to columns with identical IDs.}

\item{func}{the function to be applied}
}
\description{
A more general version of avereps
}
\details{
Just like avereps averages the expression value for each sample for
genes that have the same ID, \code{repsapply} applies any arbitrary
function.
}
