% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocplot.R
\name{rf2pr}
\alias{rf2pr}
\title{Convert random forest to format suitable for rocplot()}
\usage{
rf2pr(rf)
}
\arguments{
\item{rf}{An object returned by randomForest()}

\item{reality}{A vector with actual class assignments}
}
\description{
Convert random forest to format suitable for rocplot()
}
\details{
Convert the object returned by randomForest() to a format suitable for
the function rocplot()
}
