% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2csv.R
\name{rg2csv}
\alias{rg2csv}
\title{Save an RG object as CSV files}
\usage{
rg2csv(rg, stem = "export")
}
\arguments{
\item{rg}{RG object from limma}

\item{stem}{File stem to save to}
}
\description{
Save an RG object as CSV files
}
\details{
Save an RG object as CSV files. Each of the components E, genes and
targets will be save as a separate CSV file starting with \code{stem}.
File names will be the stem + underscore + qualifier (data for E, features
for genes and samples for targets) + .csv.
}
