% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocplot.R
\name{rocformat}
\alias{rocformat}
\title{Format AUC from roc() function}
\usage{
rocformat(x)
}
\arguments{
\item{x}{object returned by roc()}
}
\description{
Format AUC from roc() function
}
\details{
Format AUC from roc() function
}
