% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confuMat.R, R/rocplot.R
\name{rocplot}
\alias{rocplot}
\title{Calculate and show ROC curve}
\usage{
rocplot(ret, title = "ROC", file = NULL, confmat = TRUE, ci = TRUE,
  positive.name = NULL, add = FALSE, ...)

rocplot(ret, title = "ROC", file = NULL, confmat = TRUE, ci = TRUE,
  positive.name = NULL, add = FALSE, ...)
}
\arguments{
\item{reality}{a vector of responses (two possible values)}

\item{predictor}{a vector predicting the response (numeric)}

\item{decision}{a vector of predicted responses (same values as response)}
}
\value{
\code{rocplot} invisibly returns the return value of the \code{roc} call

whatever the roc() function returns
}
\description{
Calculate and show ROC curve

ROC figure using pROC package
}
\details{
Calculate and show ROC curve

This is a wrapper around the \code{roc} function from the pROC package.
It plots the ROC curve, but additionally it also can add information on the
confidence interval and a confusion matrix.
}
