% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{rocseries_boxplot}
\alias{rocseries_boxplot}
\title{plot a series of ROC prediction objects as boxplot}
\usage{
rocseries_boxplot(x, xlim = c(0.25, 1), rcol = "#ccddcc", bty = "n",
  mm = 0.4, ...)
}
\arguments{
\item{x}{a series of roc predictions (a list)}

\item{xlim}{xlim for the plot}
}
\description{
plot a series of ROC prediction objects as boxplot
}
