% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{rocseries_table}
\alias{rocseries_table}
\title{Return a table with AUC for each prediction in a series}
\usage{
rocseries_table(x, method = "fdr")
}
\description{
based on prediction objects, returns a table with auc, ci and p-value
method: method to adjust for multiple testing
}
