% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSPIA.R
\name{runSPIA}
\alias{runSPIA}
\title{wrapper for the spia function.}
\usage{
runSPIA(fit, coef = 1, organism = "mmu", p.value = 0.05, lfc = 1,
  eid.field = "EID")
}
\description{
wrapper for the spia function.
}
\details{
wrapper for the spia function.
}
