% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.xlsx.R
\name{save.xlsx}
\alias{save.xlsx}
\title{Save multiple objects to an excel file}
\usage{
save.xlsx(file, objects, names = NULL)
}
\description{
Save multiple objects to an excel file
}
\details{
Save multiple objects to an excel file
}
