% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importexport.R
\name{saveAsCLS}
\alias{saveAsCLS}
\title{Write pheno data as braindead CLS format}
\usage{
saveAsCLS(groups, FileName)
}
\arguments{
\item{groups}{A factor}

\item{FileName}{name of the file to write to}
}
\description{
Write pheno data as braindead CLS format
}
\details{
Write pheno data as braindead CLS format.
See http://www.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats
}
