% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importexport.R
\name{saveAsGCT}
\alias{saveAsGCT}
\title{Write expression data in the braindead GCT format}
\usage{
saveAsGCT(E, Name, FileName, Description = NULL, sampleNames = NULL)
}
\arguments{
\item{E}{a matrix with the data}

\item{Name}{A vector with probe or gene names}

\item{FileName}{name of the file to write to}

\item{Description}{Optional description vector}

\item{sampleNames}{unique names of the samples}
}
\description{
Write expression data in the braindead GCT format
}
\details{
Write expression data in the braindead GCT format. See http://www.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats
}
