% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{sdbar}
\alias{sdbar}
\title{Draw whiskers on a plot}
\usage{
sdbar(x, y, w, h)
}
\arguments{
\item{x, }{y: position of the whisker midpoint}

\item{w, }{h: half of the total height and width of the whisker}
}
\description{
Draw whiskers on a plot
}
\details{
Draw whiskers at specified positions with given heights and breadth on a plot
}
