% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setColnames.R
\name{setColnames}
\alias{setColnames}
\alias{setRownames}
\title{Return the object with colnames set}
\usage{
setColnames(x, colnames = NULL)

setRownames(x, rownames = NULL)
}
\value{
the same object with colnames set to colnames
}
\description{
Return the object with colnames set
}
\details{
Return the object with colnames set. This is equivalent to setNames, but
for objects that have columns like data frames and matrices.
}
