% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{\%diff\%}
\alias{\%diff\%}
\alias{\%allin\%}
\alias{\%sum\%}
\alias{\%eq\%}
\alias{\%notin\%}
\alias{\%and\%}
\title{Set difference operator}
\usage{
x \%diff\% y

x \%allin\% y

x \%sum\% y

x \%eq\% y

x \%notin\% y

x \%and\% y
}
\arguments{
\item{x, y}{a vector}
}
\description{
Set difference operator

All contained in operator

Set sum operator

Set equal operator

Set "not in" operator

Set intersection operator
}
