% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setEdgeAttr.R
\name{setEdgeAttr}
\alias{setEdgeAttr}
\title{set Edge attributes in an Rgraphviz graph}
\usage{
setEdgeAttr(graph, attribute, value, ID1, ID2)
}
\arguments{
\item{graph}{object of class Ragraph}

\item{attribute}{name of the attribute to set}

\item{value}{vector of values for the attribute}

\item{ID1, }{ID2 vectors with node names defining edges}
}
\value{
the modified graph object with attributes set
}
\description{
set Edge attributes in an Rgraphviz graph
}
\details{
set Edge attributes in an Rgraphviz graph
}
