% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{l}
\alias{l}
\alias{p}
\alias{h}
\alias{msd}
\alias{ss}
\alias{write.col}
\title{Shortcuts for common functions}
\usage{
l(...)

p(...)

h(...)

msd(x)

ss(x)

write.col(x, col.names = F, row.names = F, quote = F, ...)
}
\arguments{
\item{...}{further parameters to be passed to other functions.}

\item{x}{a numeric vector}
}
\description{
Shortcuts for common functions
}
\details{
\code{p}, \code{h} and \code{l} are simple shortcuts for \code{paste},
\code{head} and \code{length} functions, respectively.

\code{msd} prints the mean and standard deviation of a numeric vector
and returns them invisibly in a named numeric vector.

\code{ss} returns a sum of squares of a numeric vector

\code{write.col} wrapper for \code{write.table} for saving a single
column without column or rownames and quotes.
}
