% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showgene.R
\name{showgene}
\alias{showgene}
\title{A combined beeswarm / boxplot}
\usage{
showgene(data, group, main = "", pch = 19, ylim = NULL, xlab = "",
  ylab = "log2 expression", las = 2, pwcol = NULL, col = NULL,
  bty = "n", signboxes = NULL, start.at.zero = FALSE,
  comparisons = NULL, test.func = wilcox.test, ...)
}
\description{
A combined beeswarm / boxplot
}
\details{
A combined beeswarm / boxplot
}
