% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{showGraph}
\alias{showGraph}
\title{Call RCytoscape functions to display a graph}
\usage{
showGraph(graph, name = "foo2", overwriteWindow = TRUE,
  display = TRUE, layout.name = "force-directed")
}
\arguments{
\item{graph}{a graphNEL object}

\item{name}{name for the window}

\item{overwriteWindow}{whether to overwrite the window}

\item{display}{whether to display graph immediately}

\item{layout}{layout to use}
}
\description{
Call RCytoscape functions to display a graph
}
\details{
Call RCytoscape functions to display a graph
}
