% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleGOhyper.R
\name{simpleGOhyper}
\alias{simpleGOhyper}
\title{simple alternative to hyperGTest}
\usage{
simpleGOhyper(categories, sign)
}
\arguments{
\item{categories}{Vector of category assignments for the data}

\item{sign}{Either a logical vector of the same length as categories
indicating whether the given entity is significant, or a vector of
indices indicating which entities are significant.}
}
\description{
simple alternative to hyperGTest
}
