% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothPalette.R
\name{smoothPalette}
\alias{smoothPalette}
\title{Create a gradient corresponding to a numeric vector}
\usage{
smoothPalette(x, max = NULL, min = NULL, palfunc = NULL, n = 10,
  symmetrical = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{max}{value which corresponds to the right-hand side of the color
gradient}

\item{min}{value which corresponds to the left-hand side of the color}

\item{palfunc}{palette function as returned by colorRampPalette. By
default, it is a function generating a gradient from blue through white to
red.}

\item{n}{how many steps in the gradient}

\item{symmetrical}{whether the palette should be symmetrical around 0
(with default palette, that means that "0" corresponds to the color "white"}
}
\description{
Create a gradient corresponding to a numeric vector
}
\details{
This function "converts" numbers to colors. It generates a gradient based on a particular palette
function in which each color corresponds to a value or value range in a
numeric vector. The lowest value (or the value equal to min, if
specified) corresponds to one extreme of the gradient, and the highest (or
the value equal to max, if specified) to the other extreme of the gradient.
}
