% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortby.R
\name{sortby}
\alias{sortby}
\alias{rsortby}
\alias{rsortabs}
\alias{sortabs}
\title{Shortcuts for sorting data frames}
\usage{
sortby(df, key, decreasing = F, absolute = F)

rsortby(df, key, absolute = F)

rsortabs(df, key)

sortabs(df, key, decreasing = F)
}
\arguments{
\item{df}{a data frame}

\item{key}{column for sorting the data frame}

\item{abs}{whether sort should be by absolute numeric values}
}
\value{
All functions return a sorted data frame.
}
\description{
Shortcuts for sorting data frames
}
\details{
\code{sortby} sorts a data frame by column \code{key}. If
\code{absolute} is TRUE, then the numeric values in column \code{key} will
be first turned to absolute values with \code{abs()}.

\code{rsortby} is the same as \code{sortby(..., decreasing=T)}.

\code{sortabs} and \code{rsortabs} are the same as \code{sortby} and
\code{rsortby} with parameter \code{abs=TRUE}.
}
