% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_cache.R
\name{static_cache}
\alias{static_cache}
\title{Create a static cache at the specific path}
\usage{
static_cache(path, name, exprs)
}
\arguments{
\item{path}{the path to the object}

\item{name}{name of the object to save}

\item{exprs}{expression to generate the object if cache file does not exist}
}
\description{
Create a static cache at the specific path
}
