% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topTableAll.R
\name{topTableAll}
\alias{topTableAll}
\title{Create a topTable for all coefficients in a limma fit object}
\usage{
topTableAll(fit, contrasts = NULL, genelist = NULL, idcoll = NULL,
  sign.fc.col = FALSE, sig.threshold = 1e-05, msd = FALSE,
  ci = FALSE, rawp = FALSE)
}
\value{
A data frame with gene meta data and gene fold changes /
q-values. For each contrast "N", columns "logFC.N" and "qval.N" are
created (and, optionally, columns "logFC.s.N" and "msd.N").
}
\description{
Create a topTable for all coefficients in a limma fit object
}
\details{
For every contrast in \code{fit}, the log fold change and q-value
(p-value corrected for multiple testing) are calculated with the limma
function \code{topTable}. These columns are joined to common fields (such
as probe ID and gene meta data). topTableAll does not sort the data from
the fit object.
}
