% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttplot.R
\name{ttplot}
\alias{ttplot}
\title{beeswarm for a top table (result of topTableAll) plot}
\usage{
ttplot(tt, qval = 0.05, ...)
}
\arguments{
\item{tt}{an object returned by topTableAll}

\item{qval}{q-value threshold for coloring data points}
}
\description{
beeswarm for a top table (result of topTableAll) plot
}
\details{
beeswarm for a top table (result of topTableAll) plot
}
