% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttsummary.R
\name{ttsummary}
\alias{ttsummary}
\title{Print a summary of a top table.}
\usage{
ttsummary(tt, c = NULL, p.val = 0.01, lfc = 1)
}
\description{
Print a summary of a top table.
}
\details{
Print a summary of a top table.
}
