% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.sign.R
\name{whatabout}
\alias{whatabout}
\title{Inteligently find significant coefficients for a gene}
\usage{
whatabout(name, tt, qval = 0.05, lfc = 1, ...)
}
\description{
Inteligently find significant coefficients for a gene
}
