% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xclip.R
\name{xclip}
\alias{xclip}
\title{Copy a few history lines to the X primary clipboard}
\usage{
xclip(n = 5)
}
\arguments{
\item{n}{number of lines to copy, not counting the current xclip command}
}
\description{
Copy a few history lines to the X primary clipboard
}
\details{
Copy a few history lines to the X primary clipboard. This will probably
work only on X-window systems like Ubuntu, and requires the xclip
program to be installed.
}
