% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{xpredict.nvar}
\alias{xpredict.nvar}
\title{Determine optimal number of variables}
\usage{
xpredict.nvar(x, y, x.response, y.response, rep = 100, nvar = c(2, 3,
  4, 5, 6, 7, 8, 9, 10, 15, 20, 25, 30, 40, 50, 75, 100, 125))
}
\arguments{
\item{x, y}{data matrix, rows=variables, cols=samples. row names matter}

\item{x.response, y.response}{vector of responses}

\item{rep}{number of replicates for each number of variables}

\item{nvar}{number of variables to test}
}
\description{
like xpredict, but constructs a secondary, reduced model and reports
only AUC for each of the 100 replicates
}
\seealso{
xpredict
}
