% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{xpredict}
\alias{xpredict}
\title{Cross predict from one data set to another}
\usage{
xpredict(x, y, x.response, y.response, nvar = NULL, cleanx = FALSE,
  cleany = FALSE, cutoff.method = "MaxProdSpSe", return.list = FALSE,
  rf = NULL)
}
\arguments{
\item{x, y}{data matrix, rows=variables, cols=samples. row names matter}

\item{x.response, y.response}{vector of responses}

\item{nvar}{If not NULL, build a secondary model with nvar variables}

\item{cleanx}{drop samples from x if they are absent in y}

\item{cleany}{drop samples from y if they are absent in x}

\item{cutoff.method}{See "methods" from optimal.cutpoints}

\item{return.list}{if TRUE, returns a list with all additional information rather than a data frame}

\item{rf}{if rf model provided, it will be used instead of creating a new one}

\item{rep}{number of replicates for each number of variables}
}
\description{
Given two matrices, train the model on the first, apply to the second
}
